cleaned up code base, removed unused prototypes, code now compatible with -Wall -Werrorand -Wextra

+++ src/json_parser.h
@@ -8,8 +8,9 @@
 
 JsonNode* json_parse(char* buffer, ptrdiff_t length);
 JsonNode* json_parseFile(char* path);
+
 JsonNode* json_property(JsonNode*, char*);
-JsonNode* json_index(JsonNode*, int);
+JsonNode* json_index(JsonNode*, size_t);
 JsonNode* json_get(JsonNode*, ptrdiff_t, ...);
 
 #endif // JSON4C_PARSER
